initialization of objects for testing purpose.

+++ src/main/java/com/cancun/hotel/cancunhotel/domain/Room.java
@@ -18,7 +18,7 @@ public class Room {
     private Long id;
 
     @Column
-    private Long roomNumber;
+    private Integer roomNumber;
 
     @Column
     private String description;
